bullets v2
Took 45 minutes
