+++ src/const/towers.py
@@ -155,7 +155,7 @@ class TOWERS:
         "attack_cooldown_sec": 1,
         "gold_cost": 100,
         "research_cost": 0,
-        "radius": 5,
+        "radius": 50,
         "radius_splash": 0,
         "damage_air": 0,
         "damage_ground": 50,
