+++ src/views/__init__.py
@@ -1,1 +1,1 @@
-from .bullet import *
+from .map import *
